Altered main slightly to align with part 1 requirements
Added a while loop that calls lexer until we evaluate the entire file. Runs 100+ times depending on the input, but once it hits the end of the actual source_file, it should quickly end anyway.
Since we only have one token at a time, we could actually change everything to single struct rather than vector of Token, but I kept it the way it is so I didn't have to fix anything :)