Reorganization and a more complete version
As the headline says, this version removes the unecessary files and adds more to the syntax_rules.h/cpp.

As I understand it (and AI), what will happen when we finish this program is that main will call the Rat25S function, which will iterate through OptFunctionDefinitions, OptDeclarationList, and StatementList, which in turn recursively iterate through the rest of the productions  and matching the appropriate rules until we finish the source code.

I havent added the ability to output to a file yet for any of the rules, and most of them are not finished, but this is what I've got for now and I'll keep working on this tonight and tomorrow.