Attempt at backpatch and while/if code
The instructions are right in the while loop, only issue is where the instructions are and the patchback. If you run the program with the test code in prof's partial solutions, you'll see that the second PUSHM for the relational happens after the relational in our table. Not sure if that could be fixed if we swap the place of the PUSHM instruction and relational. Assign works fine though, at least it works with his test case and in the while test case.

Also ran into an issue where JUMP jumps to 1 instruction before the LABEL if you generate the label after the instruction address is assigned, and breaks the next PUSHM if you generate the LABEL before the instruction address assignment so I'm looking into that too.