Backpatch working
I got the backpatch to work. Couldnt figure out why it wouldnt work until I realized that the prof is using an array and I chose vectors, so I was never incrementing this "instrAdd" value since I would simply push_back all my elements into the vector so instrAdd was always 0.
I replaced most of the instrAdd with "instrTable.size()" so for now the issue is fixed.